feature/CU-8678qfwpw: deletd unsued script, moved all the scripts into directory scripts
