+++ scripts/get-current-date.py
